feat: Refactor TodoService into modular services
- Refactored `TodoService` to improve modularity and maintainability by delegating responsibilities to new services:
  - Added `TodoPlanningService` for file discovery and todo mapping.
  - Added `TodoTransformationService` for todo processing and duplicate detection.
  - Added `TodoPersistenceService` for persisting results and reporting.
- Introduced `TodoProcessingResult` struct to encapsulate processing results.
- Removed redundant methods from `TodoService` and updated it to use the new services.
- Added copyright and license headers to all modified files.
- Improved code readability, testability, and adherence to the Single Responsibility Principle.
